backend/requirements.txt
fastapi
uvicorn
pillow
python-multipart
sqlalchemy
passlib[bcrypt]
pyjwt
requests

ultralytics
torch
torchvision
